feat: add middle-click to remove files from context. (#1148)
* add middle-click to remove files from context.

Previously when you wanted to remove files which were previously added to the chat context, you had to click the little "x" icon one by one. When you had several files to remove, this got tedious and difficult, especially since the file box varies in width, so the "x" is not in the same space.

This change makes this operation easier by allowing you to middle-click anywhere on the file box to remove it.

* fix: mouse actions when clicked outside of the tag label

---------

Co-authored-by: Carl-Robert Linnupuu <carlrobertoh@gmail.com>